Some clarifications on NEF references
There are three things to be discussed, which are actually
rather separate. The NEF is the core modelling theory, which
is made up of the three principles of representation, transformation,
and dynamics. The SPA uses the NEF to do symbol-like processing by
representing high-level information with unit vectors, and combines
and compresses that information using the superposition and convolution
operators, respectively. Nengo is a piece of software that is used
to create and simulate models using the NEF and the SPA.

I attempted to use the correct concept in all cases, and added
citations to Chris's NEF book and my paper on Nengo 2.